+++ apps/dev-launchpad/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @stackframe/dev-launchpad
 
+## 2.7.2
+
 ## 2.7.1
 
 ## 2.7.0