+++ apps/oauth-mock-server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stackframe/oauth-mock-server
 
+## 2.7.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.2
+
 ## 2.7.1
 
 ### Patch Changes