+++ examples/cjs-test/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @stackframe/example-cjs-test
 
+## 2.7.2
+
+### Patch Changes
+
+- @stackframe/stack@2.7.2
+
 ## 2.7.1
 
 ### Patch Changes