+++ packages/stack/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stackframe/stack
 
+## 2.7.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.2
+  - @stackframe/stack-ui@2.7.2
+  - @stackframe/stack-sc@2.7.2
+
 ## 2.7.1
 
 ### Patch Changes